+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# v0.1608.1 (unreleased)
+
+## Additions
+
+  * `&Endpoint` now has an `impl Endpoint`
+  * `&Pageable` now has an `impl Pageable`
+  * `api::paged` now supports `into_iter()` and `into_iter_async()` methods.
+
+## Breaking changes
+
+  * `api::paged` now requires that the item type be `'static` (shouldn't break
+    anything).
+  * `api::Paged::into_iter` now returns a `LazilyPagedIter` with `&'a E` rather
+    than `E` (shouldn't break anything).
+
 # v0.1608.0
 
 ## Additions